# src/main/java/io/swagger/api/PipelineApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.7).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.ModelApiResponse;
import io.swagger.model.Pipeline;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-07-17T08:54:48.546Z")

@Api(value = "pipeline", description = "the pipeline API")
public interface PipelineApi {

    @ApiOperation(value = "Create pipeline", nickname = "createPipeline", notes = "This can only be done by the logged in user.", response = ModelApiResponse.class, tags={ "pipeline", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = ModelApiResponse.class) })
    @RequestMapping(value = "/pipeline",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<ModelApiResponse> createPipeline(@ApiParam(value = "Created pipeline" ,required=true )  @Valid @RequestBody Pipeline body);


    @ApiOperation(value = "Get pipeline", nickname = "getPipelineById", notes = "", response = Pipeline.class, tags={ "pipeline", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Pipeline.class),
        @ApiResponse(code = 400, message = "Invalid pipelineId supplied"),
        @ApiResponse(code = 404, message = "Pipeline not found") })
    @RequestMapping(value = "/pipeline/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Pipeline> getPipelineById(@ApiParam(value = "The pipeline that needs to be fetched.",required=true) @PathVariable("id") String id);

}